+++ pkg/clickhouse/clickhouse.go
@@ -7,6 +7,7 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"github.com/AlexAkulov/clickhouse-backup/pkg/common"
 	"net/url"
 	"os"
 	"path"
@@ -204,25 +205,39 @@ func (ch *ClickHouse) getDisksFromSystemSettings(ctx context.Context) ([]Disk, e
 	case <-ctx.Done():
 		return nil, ctx.Err()
 	default:
-		var result []struct {
-			MetadataPath string `db:"metadata_path"`
-		}
-		query := "SELECT metadata_path FROM system.tables WHERE database = 'system' AND metadata_path!='' LIMIT 1;"
-		if err := ch.SelectContext(ctx, &result, query); err != nil {
+		metadataPath, err := ch.getMetadataPath(ctx)
+		if err != nil {
 			return nil, err
 		}
-		if len(result) == 0 {
-			return nil, fmt.Errorf("can't get metadata_path from system.tables")
-		}
-		metadataPath := result[0].MetadataPath
 		dataPathArray := strings.Split(metadataPath, "/")
-		clickhouseData := path.Join(dataPathArray[:len(dataPathArray)-3]...)
+		clickhouseData := path.Join(dataPathArray[:len(dataPathArray)-1]...)
 		return []Disk{{
 			Name: "default",
 			Path: path.Join("/", clickhouseData),
 			Type: "local",
 		}}, nil
 	}
+}
+
+func (ch *ClickHouse) getMetadataPath(ctx context.Context) (string, error) {
+	var result []struct {
+		MetadataPath string `db:"metadata_path"`
+	}
+	query := "SELECT metadata_path FROM system.tables WHERE database = 'system' AND metadata_path!='' LIMIT 1;"
+	if err := ch.SelectContext(ctx, &result, query); err != nil {
+		return "", err
+	}
+	if len(result) == 0 {
+		return "", fmt.Errorf("can't get metadata_path from system.tables")
+	}
+	metadataPath := strings.Split(result[0].MetadataPath, "/")
+	if strings.Contains(result[0].MetadataPath, "/store/") {
+		result[0].MetadataPath = path.Join(metadataPath[:len(metadataPath)-4]...)
+		result[0].MetadataPath = path.Join(result[0].MetadataPath, "metadata")
+	} else {
+		result[0].MetadataPath = path.Join(metadataPath[:len(metadataPath)-2]...)
+	}
+	return path.Join("/", result[0].MetadataPath), nil
 }
 
 func (ch *ClickHouse) getDisksFromSystemDisks(ctx context.Context) ([]Disk, error) {
@@ -264,8 +279,11 @@ func (ch *ClickHouse) Close() {
 func (ch *ClickHouse) GetTables(ctx context.Context, tablePattern string) ([]Table, error) {
 	var err error
 	tables := make([]Table, 0)
-	isUUIDPresent := make([]int, 0)
-	if err = ch.SelectContext(ctx, &isUUIDPresent, "SELECT count() FROM system.settings WHERE name = 'show_table_uuid_in_table_create_query_if_not_nil'"); err != nil {
+	settings := map[string]bool{
+		"show_table_uuid_in_table_create_query_if_not_nil": false,
+		"display_secrets_in_show_and_select":               false,
+	}
+	if settings, err = ch.CheckSettingsExists(ctx, settings); err != nil {
 		return nil, err
 	}
 	skipDatabases := make([]string, 0)
@@ -273,11 +291,15 @@ func (ch *ClickHouse) GetTables(ctx context.Context, tablePattern string) ([]Tab
 	if err = ch.SelectContext(ctx, &skipDatabases, "SELECT name FROM system.databases WHERE engine IN ('MySQL','PostgreSQL','MaterializedPostgreSQL')"); err != nil {
 		return nil, err
 	}
-	allTablesSQL, err := ch.prepareAllTablesSQL(ctx, tablePattern, err, skipDatabases, isUUIDPresent)
+	allTablesSQL, err := ch.prepareAllTablesSQL(ctx, tablePattern, err, skipDatabases, settings)
 	if err != nil {
 		return nil, err
 	}
 	if err = ch.StructSelectContext(ctx, &tables, allTablesSQL); err != nil {
+		return nil, err
+	}
+	metadataPath, err := ch.getMetadataPath(ctx)
+	if err != nil {
 		return nil, err
 	}
 	for i, t := range tables {
@@ -294,7 +316,7 @@ func (ch *ClickHouse) GetTables(ctx context.Context, tablePattern string) ([]Tab
 			tables[i] = t
 			continue
 		}
-		tables[i] = ch.fixVariousVersions(t)
+		tables[i] = ch.fixVariousVersions(t, metadataPath)
 	}
 	if len(tables) == 0 {
 		return tables, nil
@@ -312,7 +334,7 @@ func (ch *ClickHouse) GetTables(ctx context.Context, tablePattern string) ([]Tab
 	return tables, nil
 }
 
-func (ch *ClickHouse) prepareAllTablesSQL(ctx context.Context, tablePattern string, err error, skipDatabases []string, isUUIDPresent []int) (string, error) {
+func (ch *ClickHouse) prepareAllTablesSQL(ctx context.Context, tablePattern string, err error, skipDatabases []string, settings map[string]bool) (string, error) {
 	isSystemTablesFieldPresent := make([]IsSystemTablesFieldPresent, 0)
 	isFieldPresentSQL := `
 		SELECT 
@@ -356,10 +378,31 @@ func (ch *ClickHouse) prepareAllTablesSQL(ctx context.Context, tablePattern stri
 	if len(isSystemTablesFieldPresent) > 0 && isSystemTablesFieldPresent[0].IsTotalBytesPresent > 0 {
 		allTablesSQL += " ORDER BY total_bytes DESC"
 	}
-	if len(isUUIDPresent) > 0 && isUUIDPresent[0] > 0 {
-		allTablesSQL += " SETTINGS show_table_uuid_in_table_create_query_if_not_nil=1"
-	}
+	allTablesSQL = ch.addSettingsSQL(allTablesSQL, settings)
 	return allTablesSQL, nil
+}
+
+func (ch *ClickHouse) addSettingsSQL(inputSQL string, settings map[string]bool) string {
+	maxI := 0
+	for _, v := range settings {
+		if v {
+			maxI += 1
+		}
+	}
+	i := 0
+	for k, v := range settings {
+		if v {
+			if i == 0 {
+				inputSQL += " SETTINGS "
+			}
+			inputSQL += k + "=1"
+			if i < maxI-1 {
+				inputSQL += ", "
+			}
+			i++
+		}
+	}
+	return inputSQL
 }
 
 // GetDatabases - return slice of all non system databases for backup
@@ -381,6 +424,10 @@ func (ch *ClickHouse) GetDatabases(ctx context.Context, cfg *config.Config, tabl
 		if strings.HasSuffix(pattern, ".*") {
 			bypassDatabases = append(bypassDatabases, strings.TrimSuffix(pattern, ".*"))
 		}
+	}
+	metadataPath, err := ch.getMetadataPath(ctx)
+	if err != nil {
+		return nil, err
 	}
 	select {
 	case <-ctx.Done():
@@ -410,6 +457,15 @@ func (ch *ClickHouse) GetDatabases(ctx context.Context, cfg *config.Config, tabl
 				ch.Log.Warnf("can't get create database query: %v", err)
 				allDatabases[i].Query = fmt.Sprintf("CREATE DATABASE `%s` ENGINE = %s", db.Name, db.Engine)
 			} else {
+				// 23.3+ masked secrets https://github.com/AlexAkulov/clickhouse-backup/issues/640
+				if strings.Contains(result[0], "'[HIDDEN]'") {
+					if attachSQL, err := os.ReadFile(path.Join(metadataPath, common.TablePathEncode(db.Name)+".sql")); err != nil {
+						return nil, err
+					} else {
+						result[0] = strings.Replace(string(attachSQL), "ATTACH", "CREATE", 1)
+						result[0] = strings.Replace(result[0], " _ ", " `"+db.Name+"` ", 1)
+					}
+				}
 				allDatabases[i].Query = result[0]
 			}
 		}
@@ -431,7 +487,7 @@ func (ch *ClickHouse) getTableSizeFromParts(ctx context.Context, table Table) ui
 	return 0
 }
 
-func (ch *ClickHouse) fixVariousVersions(t Table) Table {
+func (ch *ClickHouse) fixVariousVersions(t Table, metadataPath string) Table {
 	// versions before 19.15 contain data_path in a different column
 	if t.DataPath != "" {
 		t.DataPaths = []string{t.DataPath}
@@ -449,6 +505,16 @@ func (ch *ClickHouse) fixVariousVersions(t Table) Table {
 		t.CreateTableQuery = strings.Replace(
 			t.CreateTableQuery, "CREATE MATERIALIZED VIEW", "ATTACH MATERIALIZED VIEW", 1,
 		)
+	}
+	// 23.3+ masked secrets https://github.com/AlexAkulov/clickhouse-backup/issues/640
+	if strings.Contains(t.CreateTableQuery, "'[HIDDEN]'") {
+		tableSQLPath := path.Join(metadataPath, common.TablePathEncode(t.Database), common.TablePathEncode(t.Name)+".sql")
+		if attachSQL, err := os.ReadFile(tableSQLPath); err != nil {
+			ch.Log.Warnf("can't read %s: %v", tableSQLPath, err)
+		} else {
+			t.CreateTableQuery = strings.Replace(string(attachSQL), "ATTACH", "CREATE", 1)
+			t.CreateTableQuery = strings.Replace(t.CreateTableQuery, " _ ", " `"+t.Database+"`.`"+t.Name+"` ", 1)
+		}
 	}
 	return t
 }
@@ -1109,4 +1175,26 @@ func (ch *ClickHouse) CheckSystemPartsColumns(ctx context.Context, table *Table)
 		return fmt.Errorf("`%s`.`%s` have inconsistent data types for active data part in system.parts_columns", table.Database, table.Name)
 	}
 	return nil
+}
+
+func (ch *ClickHouse) CheckSettingsExists(ctx context.Context, settings map[string]bool) (map[string]bool, error) {
+	isSettingsPresent := make([]*struct {
+		Name      string `db:"name"`
+		IsPresent int    `db:"is_present"`
+	}, 0)
+	queryStr := "SELECT name, count(*) as is_present FROM system.settings WHERE name IN ("
+	args := make([]interface{}, 0, len(settings))
+	for k := range settings {
+		queryStr += "?, "
+		args = append(args, k)
+	}
+	queryStr = queryStr[:len(queryStr)-2]
+	queryStr += ") GROUP BY name"
+	if err := ch.SelectContext(ctx, &isSettingsPresent, queryStr, args...); err != nil {
+		return nil, err
+	}
+	for _, item := range isSettingsPresent {
+		settings[item.Name] = item.IsPresent > 0
+	}
+	return settings, nil
 }